Fix presubmit tests
Our Android.bp and TEST_MAPPING files were incorrect, so tests were not running
on presubmit. This change corrects this.

In addition, several tests were failing because we do not include sample trace
data in the repo. These tests are now skipped by introducing a @NeedsSampleData
annotation that is used to filter these tests out.

Change-Id: Ie7172a01bb9cf2fa6142fb0f70c1a698fa2dbe15